// pro/src/apiClient/v1/models/CollectiveAdditionalFeeModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CollectiveAdditionalFeeType } from './CollectiveAdditionalFeeType';
export type CollectiveAdditionalFeeModel = {
  amount: number;
  label: (string | null);
  type: CollectiveAdditionalFeeType;
};